requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
pillow
aiofiles
pydantic-settings
python-magic